- Added Key Listener, to listen for specific key presses - Rotation can now be at any angle, not just whole numbers - Added the Unsafe Rotations keybind to rotate any object at any angle, rather than predefined rotations - Coloured objects can now be solid or hazard colliders - Fixed an issue where players needed to have the edit toggle enabled in order for levels to refresh properly over multiplayer - Added the Watcher Knights - The Trigger Zone can now be set to use custom width and height values

+++ WorldEditor/Architect/Content/Elements/AbstractPackElement.cs
@@ -6,7 +6,7 @@ namespace Architect.Content.Elements;
 
 public abstract class AbstractPackElement
 {
-    public abstract GameObject GetPrefab(bool flipped, int rotation);
+    public abstract GameObject GetPrefab(bool flipped, float rotation);
 
     private readonly string _name;
     private readonly string _category;
@@ -123,5 +123,5 @@ public abstract class AbstractPackElement
         return false;
     }
 
-    public virtual void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale) { }
+    public virtual void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale) { }
 }
