+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/ColosseumPackElement.cs
@@ -19,7 +19,7 @@ internal class ColosseumPackElement : InternalPackElement
         WithBroadcasterGroup(BroadcasterGroup.Enemies);
     }
 
-    public override GameObject GetPrefab(bool flipped, int rotation)
+    public override GameObject GetPrefab(bool flipped, float rotation)
     {
         return _gameObject;
     }
