+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/ElderBaldurElement.cs
@@ -21,7 +21,7 @@ internal class ElderBaldurElement : InternalPackElement
         return true;
     }
 
-    public override GameObject GetPrefab(bool flipped, int rotation)
+    public override GameObject GetPrefab(bool flipped, float rotation)
     {
         return _gameObject;
     }
@@ -36,7 +36,7 @@ internal class ElderBaldurElement : InternalPackElement
         _gameObject = preloads["Crossroads_ShamanTemple"]["Battle Scene/Blocker"];
     }
 
-    public override void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale)
+    public override void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale)
     {
         gameObject.transform.SetScaleX(-_gameObject.transform.GetScaleX());
         if (flipped) return;
