+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/HopperPackElement.cs
@@ -15,7 +15,7 @@ internal sealed class HopperPackElement : GInternalPackElement
         _flip = flip;
     }
 
-    public override void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale)
+    public override void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale)
     {
         var fsm = gameObject.LocateMyFSM("Hopper");
