+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/MillibellePackElement.cs
@@ -13,7 +13,7 @@ internal class MillibellePackElement : InternalPackElement
         FlipVertical();
     }
 
-    public override GameObject GetPrefab(bool flipped, int rotation)
+    public override GameObject GetPrefab(bool flipped, float rotation)
     {
         return _gameObject;
     }
@@ -28,7 +28,7 @@ internal class MillibellePackElement : InternalPackElement
         _gameObject = preloads["Ruins_Bathhouse"]["Banker Spa NPC"];
     }
 
-    public override void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale)
+    public override void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale)
     {
         var fsm = gameObject.LocateMyFSM("Hit Around");
         fsm.DisableAction("Init", 1);
