+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/ReusableLeverElement.cs
@@ -17,7 +17,7 @@ internal class ReusableLeverElement : InternalPackElement
         WithRotationGroup(RotationGroup.Eight);
     }
 
-    public override GameObject GetPrefab(bool flipped, int rotation)
+    public override GameObject GetPrefab(bool flipped, float rotation)
     {
         return _gameObject;
     }
@@ -32,7 +32,7 @@ internal class ReusableLeverElement : InternalPackElement
         _gameObject = preloads["Ruins1_03"]["Lift Call Lever"];
     }
 
-    public override void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale)
+    public override void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale)
     {
         var fsm = gameObject.LocateMyFSM("Call Lever");
