+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/TwisterPackElement.cs
@@ -17,7 +17,7 @@ internal sealed class TwisterPackElement : GInternalPackElement
         _fsmName = fsmName;
     }
 
-    public override void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale)
+    public override void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale)
     {
         var fsm = gameObject.LocateMyFSM(_fsmName);
