+++ WorldEditor/Architect/Content/Elements/Internal/Fixers/WhitePalaceShieldGateElement.cs
@@ -16,7 +16,7 @@ internal class WhitePalaceShieldGateElement : InternalPackElement
         WithConfigGroup(ConfigGroup.BattleGate);
     }
 
-    public override GameObject GetPrefab(bool flipped, int rotation)
+    public override GameObject GetPrefab(bool flipped, float rotation)
     {
         return _gameObject;
     }
@@ -31,7 +31,7 @@ internal class WhitePalaceShieldGateElement : InternalPackElement
         _gameObject = preloads["White_Palace_20"]["Battle Scene/Gate"];
     }
 
-    public override void PostSpawn(GameObject gameObject, bool flipped, int rotation, float scale)
+    public override void PostSpawn(GameObject gameObject, bool flipped, float rotation, float scale)
     {
         var collider = gameObject.GetComponent<BoxCollider2D>();
         collider.offset = new Vector2(0.25f, 0.2165f);
