+++ WorldEditor/Architect/Content/Elements/SimplePackElement.cs
@@ -24,7 +24,7 @@ public class SimplePackElement : AbstractPackElement
         _sprite = sprite;
     }
 
-    public override GameObject GetPrefab(bool flipped, int rotation)
+    public override GameObject GetPrefab(bool flipped, float rotation)
     {
         return _obj; 
     }
